requirements.txt
bcrypt==4.3.0
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
cryptography==45.0.2
future==1.0.0
gitdb==4.0.11
gitpython==3.1.40
idna==3.4
ipaddress==1.0.23
ipython==9.2.0
jinja2==3.1.2
junos-eznc==2.7.1
lxml==5.4.0
markdown-it-py==3.0.0
markupsafe==2.1.3
mdurl==0.1.2
napalm==5.0.0
ncclient==0.6.19
netaddr==1.3.0
netutils==1.13.0
nornir==3.4.1
nornir-inspect==1.0.3
nornir-jinja2==0.2.0
nornir-napalm==0.5.0
nornir-netbox==0.3.0
nornir-netmiko==1.0.1
nornir-utils==0.2.0
ntc-templates==7.8.0
packaging==23.2
paramiko==3.5.1
pathspec==0.11.2
pbr==5.11.1
pip==23.3.1
platformdirs==3.11.0
pycparser==2.22
pyeapi==1.0.4
pygments==2.16.1
pynacl==1.5.0
pyparsing==3.2.3
pyserial==3.5
python-dotenv==1.0.0
pyyaml==6.0.2
requests==2.31.0
ruamel-yaml==0.18.3
ruamel-yaml-clib==0.2.8
scp==0.15.0
setuptools==58.1.0
six==1.17.0
smmap==5.0.1
snowballstemmer==2.2.0
stevedore==5.1.0
textfsm==1.1.3
tomli==2.0.1
transitions==0.9.2
ttp==0.9.5
ttp-templates==0.3.7
typing-extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.5
yamlordereddictloader==0.4.2
